Delete all logic in todo.js
Want to start from scratch and simplify things.
The idea is that the task are gonna be divided
in the projects, so that when todos are added
they are added to project that is active in that
moment.
Also the filters are gonna only use the filter
method to display specific tasks.